Fix int/long type mismatch for DSM _L variables in ACUmetrology thread
DSM variable names with _L suffix require long (8 bytes on 64-bit Linux),
but the intermediate local variables were declared as int (4 bytes). This
caused dsm_write to read 8 bytes from a 4-byte buffer, writing garbage in
the upper 32 bits of each DSM entry. All readers (gltmonitor, acuMonitor)
then got corrupted values for the metrology variables.

Fix: declare all local variables used with _L DSM writes as long, so the
assignment from the 32-bit metrologyData struct fields sign-extends cleanly
and dsm_write gets a properly-sized buffer.

+++ glttrack/glttrack.c
@@ -3974,12 +3974,12 @@ void *ACUmetrology() {
   short checksum;
 
   char acuErrorMessage[256],generalMetStatus[4];
-  int tilt1x,tilt1y,tilt2x,tilt2y,tilt3x,tilt3y;
-  int tilt1Temp,tilt2Temp,tilt3Temp;
-  int linearSensor1,linearSensor2,linearSensor3,linearSensor4;
-  int tiltAzCorr,tiltElCorr;
-  int linearSensorAzCorr,linearSensorElCorr;
-  int SPEMazCorr,SPEMelCorr;
+  long tilt1x,tilt1y,tilt2x,tilt2y,tilt3x,tilt3y;
+  long tilt1Temp,tilt2Temp,tilt3Temp;
+  long linearSensor1,linearSensor2,linearSensor3,linearSensor4;
+  long tiltAzCorr,tiltElCorr;
+  long linearSensorAzCorr,linearSensorElCorr;
+  long SPEMazCorr,SPEMelCorr;
   short tempSensor[41];
   
   acuCommand.stx = 0x2;
